+++ community-marketplace/web/prefab-viewer/PrefabViewer.js
@@ -3,8 +3,8 @@
  */
 import * as THREE from "three";
 import { OrbitControls } from "three/addons/controls/OrbitControls.js";
-import { loadCatalogs } from "./BlockCatalog.js?v=43";
-import { buildPrefabMesh, disposeObject3D, PREFAB_VIEWER_TRANSFORM_REV } from "./PrefabMeshBuilder.js?v=43";
+import { loadCatalogs } from "./BlockCatalog.js?v=44";
+import { buildPrefabMesh, disposeObject3D, PREFAB_VIEWER_TRANSFORM_REV } from "./PrefabMeshBuilder.js?v=44";
 
 export { PREFAB_VIEWER_TRANSFORM_REV };